@@ -1,9 +1,9 @@
 // Benchmark boilerplate generated by Benchmark
 
 import Benchmark
+import CoreGraphicsSupport
 import Foundation
 import Shapes2D
-import CoreGraphicsSupport
 
 let benchmarks = {
     Benchmark("DeCasteljau") { benchmark in
@@ -44,7 +44,6 @@ func testBernsteinPolynomal() {
     }
 }
 
-
 func testBernsteinPolynomalMatrix() {
     let curve = CubicBezierCurve(controlPoints: [[0, 0], [100, 0], [100, 100], [0, 100]])
     let controlPoints = curve.controlPoints
@@ -56,5 +55,4 @@ func testBernsteinPolynomalMatrix() {
         blackHole(xSolver.sample_cubic_matrix(t: t))
         blackHole(ySolver.sample_cubic_matrix(t: t))
     }
-
 }
