@@ -1,17 +1,9 @@
-import Algorithms
-import CoreGraphicsSupport
-import Observation
-import Sketches
+import SwiftGraphicsDemosSupport
 import SwiftUI
-import Shapes2D
 
 struct ContentView: View {
-
     @Environment(\.openWindow)
     var openWindow
-
-    @State
-    var sketch = Sketch()
 
     var body: some View {
         SketchView()
@@ -22,5 +14,3 @@ struct ContentView: View {
             }
     }
 }
-
-
