@@ -1,5 +1,4 @@
-//
-import LegacyGeometry
+import GenericGeometryBase
 
 // TODO: Reduce Storage requirements.
 public struct Accessor2D<Storage> where Storage: RandomAccessCollection & MutableCollection, Storage.Index == Int {
