@@ -1,5 +1,5 @@
 import Foundation
-import LegacyGeometry
+import GenericGeometryBase
 
 public struct Array2D<Element> {
 //    public enum Order {
@@ -11,8 +11,6 @@ public struct Array2D<Element> {
 
     public private(set) var size: IntSize
     public var flatStorage: [Element]
-
-    var debugLabel: String?
 
     public init(repeating: Element, size: IntSize) {
         self.size = size
