@@ -2,7 +2,6 @@ import SwiftUI
 
 // Treating SwiftUI.Path as part of CoreGraphics
 
-
 public extension Path {
     init(lines: [CGPoint]) {
         self.init()
@@ -12,9 +11,9 @@ public extension Path {
     // TODO: line segment
     init(lineSegments: [(CGPoint, CGPoint)]) {
         self.init()
-        lineSegments.forEach {
-            move(to: $0.0)
-            addLine(to: $0.1)
+        for lineSegment in lineSegments {
+            move(to: lineSegment.0)
+            addLine(to: lineSegment.1)
         }
     }
 
@@ -38,14 +37,13 @@ public extension Path {
         }
         self = path.strokedPath(StrokeStyle(lineWidth: width, lineCap: lineCap))
     }
-
 }
 
 public extension Path {
     init(paths: [Path]) {
         self.init()
         for path in paths {
-            self.addPath(path)
+            addPath(path)
         }
     }
 }
@@ -98,33 +96,39 @@ public extension Path {
     static func rect(x: Double = 0, y: Double = 0, w: Double, h: Double) -> Path {
         Path(CGRect(x: x, y: y, width: w, height: h))
     }
+
     static func rect(x: Double = 0, y: Double = 0, w: Double, h: Double, r: Double, style: RoundedCornerStyle = .circular) -> Path {
         Path(roundedRect: CGRect(x: x, y: y, width: w, height: h), cornerSize: CGSize(width: r, height: r), style: style)
     }
+
     static func rect(x: Double = 0, y: Double = 0, w: Double, h: Double, rx: Double, ry: Double, style: RoundedCornerStyle = .circular) -> Path {
         Path(roundedRect: CGRect(x: x, y: y, width: w, height: h), cornerSize: CGSize(width: rx, height: ry), style: style)
     }
-    
+
     static func circle(center: CGPoint, radius: Double) -> Path {
         Path(ellipseIn: CGRect(center: center, radius: radius))
     }
-    
+
     static func circle(cx: Double, cy: Double, r: Double) -> Path {
         Path(ellipseIn: CGRect(center: CGPoint(x: cx, y: cy), radius: r))
     }
-    
+
     static func ellipse(center: CGPoint, radius: CGSize) -> Path {
         Path(ellipseIn: CGRect(center: center, size: radius * 2))
     }
+
     static func ellipse(center: CGPoint, rx: Double, ry: Double) -> Path {
         ellipse(center: center, radius: CGSize(width: rx, height: ry))
     }
+
     static func ellipse(cx: Double, cy: Double, rx: Double, ry: Double) -> Path {
         ellipse(center: CGPoint(x: cx, y: cy), radius: CGSize(width: rx, height: ry))
     }
+
     static func ellipse(cx: Double, cy: Double, r: Double) -> Path {
         ellipse(center: CGPoint(x: cx, y: cy), radius: CGSize(width: r, height: r))
     }
+
     static func lines(_ lines: [CGPoint], closed: Bool = false) -> Path {
         Path { path in
             path.addLines(lines)
@@ -133,27 +137,28 @@ public extension Path {
             }
         }
     }
+
     static func line(from: CGPoint, to: CGPoint) -> Path {
         Path { path in
             path.move(to: from)
             path.addLine(to: to)
         }
     }
-    
+
     static func horizontalLine(from: CGPoint, length: Double) -> Path {
         line(from: from, to: CGPoint(x: from.x + length, y: from.y))
     }
-    
+
     static func verticalLine(from: CGPoint, length: Double) -> Path {
         line(from: from, to: CGPoint(x: from.x, y: from.y + length))
     }
-    
+
     // curve
     // arc
     // regular polygon
     // star
     // cross
-    
+
     static func cross(_ rect: CGRect) -> Path {
         Path { path in
             path.moveTo(x: rect.minX, y: rect.midY)
@@ -162,7 +167,7 @@ public extension Path {
             path.addLineTo(x: rect.midX, y: rect.maxY)
         }
     }
-    
+
     static func saltire(_ rect: CGRect) -> Path {
         Path { path in
             path.moveTo(x: rect.minX, y: rect.minY)
@@ -183,7 +188,7 @@ public extension Path {
             }
         }
     }
-    
+
     static func star(points: Int, innerRadius: Double, outerRadius: Double) -> Path {
         var path = Path()
         assert(points > 1, "Number of points should be greater than 1 for a star")
@@ -204,7 +209,7 @@ public extension Path {
     }
 }
 
-extension Path {
+public extension Path {
     func scaled(x: CGFloat, y: CGFloat) -> Path {
         let transform = CGAffineTransform(translationX: -boundingRect.midX, y: -boundingRect.midY)
             .concatenating(CGAffineTransform(scaleX: x, y: y))
@@ -249,7 +254,6 @@ public extension Path {
     }
 }
 
-
 public extension Path {
     var elements: [Element] {
         var elements: [Element] = []
@@ -259,7 +263,7 @@ public extension Path {
 
     init(elements: [Element]) {
         self = Path { path in
-            elements.forEach { element in
+            for element in elements {
                 switch element {
                 case .move(let point):
                     path.move(to: point)
@@ -281,10 +285,10 @@ public extension Path {
     struct Corners: OptionSet {
         public let rawValue: UInt8
 
-        public static let topLeft = Corners(rawValue: 0b0001)
-        public static let topRight = Corners(rawValue: 0b0010)
-        public static let bottomLeft = Corners(rawValue: 0b0100)
-        public static let bottomRight = Corners(rawValue: 0b1000)
+        public static let topLeft = Self(rawValue: 0b0001)
+        public static let topRight = Self(rawValue: 0b0010)
+        public static let bottomLeft = Self(rawValue: 0b0100)
+        public static let bottomRight = Self(rawValue: 0b1000)
 
         public static let all: Corners = [.topLeft, .topRight, .bottomLeft, .bottomRight]
 
@@ -336,4 +340,3 @@ public extension Path {
             })
     }
 }
-
