@@ -1,18 +1,18 @@
 #if os(iOS) && !targetEnvironment(simulator)
-    import ARKit
-    import Metal
+import ARKit
+import Metal
 
-    public extension MTLPrimitiveType {
-        init(_ type: ARGeometryPrimitiveType) {
-            switch type {
-            case .line:
-                self = .line
-            case .triangle:
-                self = .triangle
-            @unknown
-            default:
-                fatal(error: MetalSupportError.illegalValue)
-            }
+public extension MTLPrimitiveType {
+    init(_ type: ARGeometryPrimitiveType) {
+        switch type {
+        case .line:
+            self = .line
+        case .triangle:
+            self = .triangle
+        @unknown
+        default:
+            fatal(error: MetalSupportError.illegalValue)
         }
     }
+}
 #endif
