@@ -1,6 +1,86 @@
 import simd
+import SwiftUI
 
-// NOTE: Too much duplication here. Deprecate what isn't used.
+public protocol ProjectionProtocol: Equatable, Sendable {
+    func matrix(viewSize: SIMD2<Float>) -> simd_float4x4
+}
+
+public struct PerspectiveProjection: ProjectionProtocol {
+    public var fovy: Angle
+    public var zClip: ClosedRange<Float>
+
+    public init(fovy: Angle, zClip: ClosedRange<Float>) {
+        self.fovy = fovy
+        self.zClip = zClip
+    }
+
+    public func matrix(viewSize: SIMD2<Float>) -> simd_float4x4 {
+        let aspect = viewSize.x / viewSize.y
+        return .perspective(aspect: aspect, fovy: Float(fovy.radians), near: zClip.lowerBound, far: zClip.upperBound)
+    }
+}
+
+public struct OrthographicProjection: ProjectionProtocol {
+    public var left: Float
+    public var right: Float
+    public var bottom: Float
+    public var top: Float
+    public var near: Float
+    public var far: Float
+
+    public init(left: Float, right: Float, bottom: Float, top: Float, near: Float, far: Float) {
+        self.left = left
+        self.right = right
+        self.bottom = bottom
+        self.top = top
+        self.near = near
+        self.far = far
+    }
+
+    public func matrix(viewSize: SIMD2<Float>) -> simd_float4x4 {
+        .orthographic(left: left, right: right, bottom: bottom, top: top, near: near, far: far)
+    }
+}
+
+public enum Projection: ProjectionProtocol {
+    case matrix(simd_float4x4)
+    case perspective(PerspectiveProjection)
+    case orthographic(OrthographicProjection)
+
+    public func matrix(viewSize: SIMD2<Float>) -> simd_float4x4 {
+        switch self {
+        case .matrix(let projection):
+            projection
+        case .perspective(let projection):
+            projection.matrix(viewSize: viewSize)
+        case .orthographic(let projection):
+            projection.matrix(viewSize: viewSize)
+        }
+    }
+
+    // TODO: Use that macro
+    public enum Meta: CaseIterable {
+        case matrix
+        case perspective
+        case orthographic
+    }
+
+    public var meta: Meta {
+        switch self {
+        case .matrix:
+            .matrix
+        case .perspective:
+            .perspective
+        case .orthographic:
+            .orthographic
+        }
+    }
+}
+
+extension Projection: Sendable {
+}
+
+// TODO: Too much duplication here. Deprecate what isn't used.
 
 public extension simd_float4x4 {
     // swiftlint:disable:next function_parameter_count
@@ -12,9 +92,7 @@ public extension simd_float4x4 {
             [0, 0, 0, 1]
         ))
     }
-}
 
-public extension simd_float4x4 {
     static func perspective(aspect: Float, fovy: Float, near: Float, far: Float) -> Self {
         let yScale = 1 / tan(fovy * 0.5)
         let xScale = yScale / aspect
@@ -29,9 +107,7 @@ public extension simd_float4x4 {
 
         return simd_float4x4([P, Q, R, S])
     }
-}
 
-public extension simd_float4x4 {
     static func viewport(x: Float, y: Float, w: Float, h: Float, depth: Float) -> simd_float4x4 {
         var m = simd_float4x4.identity
         m[0][3] = x + w / 2
