@@ -1,6 +1,8 @@
 import Algorithms
 import CoreGraphics
 import Earcut
+import MetalSupport
+import Shapes2D
 import simd
 
 // swiftlint:disable force_unwrapping
@@ -24,7 +26,7 @@ public extension ExtrusionAxis {
     }
 }
 
-public extension PolygonalChain where Point == CGPoint {
+public extension PolygonalChain {
     func extrude(min: Float, max: Float, axis: ExtrusionAxis = .z) -> TrivialMesh<SimpleVertex> {
         let quads: [Quad<SimpleVertex>] = vertices.windows(ofCount: 2).reduce(into: []) { result, window in
             let from = SIMD2<Float>(x: Float(window.first!.x), y: Float(window.first!.y))
@@ -61,9 +63,9 @@ extension Quad {
     }
 }
 
-public extension Polygon where Vertex == CGPoint {
+public extension Polygon {
     func extrude(min: Float, max: Float, axis: ExtrusionAxis = .z, walls: Bool = true, topCap: Bool, bottomCap: Bool) -> TrivialMesh<SimpleVertex> {
-        let walls = walls ? PolygonalChain(polygon: self).extrude(min: min, max: max, axis: axis) : nil
+        let walls = walls ? PolygonalChain(vertices: vertices).extrude(min: min, max: max, axis: axis) : nil
         let topCap = topCap ? triangulate(z: max, transform: axis.transform) : nil
         let bottomCap = bottomCap ? triangulate(z: min, transform: axis.transform).flipped() : nil
         return TrivialMesh(merging: Array([walls, topCap, bottomCap].compacted()))
